// npm_api/src/generated/select/runtime/generateGraphqlOperation.ts
export interface GraphqlOperation {
    query: string
    variables?: { [name: string]: any }
    operationName?: string
}